clamp layer title and add tooltip
