+++ packages/geoview-core/src/core/components/common/index.ts
@@ -1,7 +1,6 @@
 export * from './focus-trap-element';
 export * from './responsive-grid';
 export * from './layer-list';
-export * from './layer-title';
 export * from './layout';
 export * from './use-footer-panel-height';
 export * from './use-lightbox';
